rm out-dated, refactor, py3 compat
