+++ ispyb/interface/acquisition.py
@@ -1,22 +1,22 @@
 import abc
+ABC = abc.ABCMeta('ABC', (object,), {'__slots__': ()}) # compatible with Python 2 *and* 3
 
-class IF(object):
-  __metaclass__ = abc.ABCMeta
+class IF(ABC):
 
   @abc.abstractmethod
   def get_data_collection_group_params(cls):
-    raise NotImplementedError('users must define this method to use this base class')
+    pass
 
   @abc.abstractmethod
   def get_data_collection_params(cls):
-    raise NotImplementedError('users must define this method to use this base class')
+    pass
 
   @abc.abstractmethod
   def upsert_data_collection_group(cls, cursor, values):
     '''Store new MX data collection group.'''
-    raise NotImplementedError('users must define this method to use this base class')
+    pass
 
   @abc.abstractmethod
   def upsert_data_collection(cls, cursor, values):
     '''Store new data collection.'''
-    raise NotImplementedError('users must define this method to use this base class')
+    pass
